pyproject.toml
[project]
name = "istat-mcp-server"
version = "0.1.0"
description = "MCP server for accessing Italian ISTAT statistical data via SDMX API"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Vincenzo Patruno", email = "patrunomeister@gmail.com" }
]
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "lxml>=5.0.0",
    "cachetools>=5.3.0",
    "diskcache>=5.6.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "duckdb>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
istat-get-concepts-cli = "istat_mcp_server.cli.get_concepts_cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
istat_mcp_server = ["resources/*.duckdb"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
